Greatly simplified loading & navigation flow
Now, we *always* listen for new content and update the page based on it.
This is how the codebase was in the past, we'd gotten away from it a bit.

So, we no longer care about what page we're on - if we detect any LE
content, we try to modify it.