Made DOM manipulation idempotent
We were adding more and more complex logic around keeping track of which
elements had been modified and which not, and I realized that it would
be vastly simpler to just make modification idempotent. That way, we can
re-run the modification code over the same element over and over and
it will render correctly every time.